createcontext to get meals in storage and remove get storage in screen Home
